Removed PartitionKeyUtil and calls to it with StringPartitioningStrategy.
