+++ hazelcast/src/main/java/com/hazelcast/partition/strategy/StringAndPartitionAwarePartitioningStrategy.java
@@ -18,7 +18,6 @@ package com.hazelcast.partition.strategy;
 
 import com.hazelcast.core.PartitionAware;
 import com.hazelcast.core.PartitioningStrategy;
-import com.hazelcast.util.PartitionKeyUtil;
 
 /**
  * @author mdogan 8/25/13
@@ -27,7 +26,7 @@ public class StringAndPartitionAwarePartitioningStrategy implements Partitioning
 
     public Object getPartitionKey(Object key) {
         if (key instanceof String) {
-            String partitionKey = PartitionKeyUtil.getPartitionKey((String) key);
+            String partitionKey = StringPartitioningStrategy.getPartitionKey((String) key);
             return partitionKey != key ? partitionKey : null;
         } else if (key instanceof PartitionAware) {
             return ((PartitionAware) key).getPartitionKey();
